feat: :sparkles: convert phase to rxjs, persist in local storage
