+++ client/package.json
@@ -30,8 +30,8 @@
     "ts-jest": "^29.0.3",
     "typescript": "~4.8.2",
     "vite": "^3.1.7",
-    "vite-tsconfig-paths": "^3.5.2",
-    "vite-svg-loader": "^3.6.0"
+    "vite-svg-loader": "^3.6.0",
+    "vite-tsconfig-paths": "^3.5.2"
   },
   "dependencies": {
     "@equinor/eds-core-react": "^0.23.0",
@@ -42,7 +42,9 @@
     "@equinor/fusion-framework-react-app": "^1.0.15",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-query": "^3.39.2",
     "react-router-dom": "^6.4.2",
+    "rxjs": "^7.5.7",
     "styled-components": "^5.3.6"
   }
 }
