--- client/packages/portal-client/src/components/phase-icons/Early.tsx
@@ -1,53 +1,0 @@
-export const Early = () => {
-  return (
-    <svg
-      width="318"
-      height="305"
-      viewBox="0 0 318 305"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        d="M98 175L48 175V167.075C71.3409 164.193 87.2757 155.003 98 145.694V175Z"
-        fill="#FF6670"
-      />
-      <path
-        d="M32.5 168C68.5 168 90 154 102.5 141.5C115 129 117 105 152.5 105C188 105 187 126.5 202 141.5C217 156.5 241.5 168 270.5 168"
-        stroke="#3D3D3D"
-        strokeWidth="4"
-        strokeLinecap="round"
-      />
-      <line
-        x1="32"
-        y1="174"
-        x2="271"
-        y2="174"
-        stroke="#3D3D3D"
-        strokeWidth="4"
-        strokeLinecap="round"
-      />
-      <line
-        x1="49"
-        y1="134"
-        x2="49"
-        y2="180"
-        stroke="#3D3D3D"
-        strokeWidth="4"
-        strokeLinecap="round"
-      />
-      <line
-        x1="99"
-        y1="134"
-        x2="99"
-        y2="180"
-        stroke="#3D3D3D"
-        strokeWidth="4"
-        strokeLinecap="round"
-      />
-      <path
-        d="M126.719 264.672C127.007 264.864 127.247 264.744 127.247 264.384V261.84C127.247 261.6 127.151 261.456 126.935 261.312L122.135 258.504L126.935 255.696C127.151 255.552 127.247 255.408 127.247 255.168V252.624C127.247 252.264 127.007 252.144 126.719 252.336L119.015 257.256C118.823 257.376 118.727 257.544 118.727 257.784V259.224C118.727 259.464 118.823 259.632 119.015 259.752L126.719 264.672ZM141.492 267C145.932 267 149.268 263.904 149.268 258.6C149.268 253.296 145.932 250.2 141.492 250.2H135.972C135.684 250.2 135.516 250.368 135.516 250.632V266.568C135.516 266.832 135.684 267 135.972 267H141.492ZM138.612 253.032H141.276C143.988 253.032 146.124 254.688 146.124 258.6C146.124 262.512 143.988 264.168 141.276 264.168H138.612V253.032ZM161.211 257.496C160.971 257.496 160.779 257.592 160.707 257.856L159.987 259.824C159.867 260.136 160.011 260.328 160.323 260.328H165.099C164.571 262.8 162.891 264.408 160.107 264.408C156.627 264.408 154.611 261.936 154.611 258.6C154.611 255.264 156.795 252.792 160.107 252.792C161.931 252.792 163.443 253.656 164.355 254.664C164.547 254.88 164.763 254.88 164.979 254.712L166.659 253.344C166.875 253.176 166.899 252.936 166.731 252.72C165.099 250.848 162.771 249.912 160.155 249.912C155.259 249.912 151.467 253.608 151.467 258.6C151.467 263.592 155.091 267.288 160.155 267.288C164.931 267.288 168.075 263.952 168.291 259.176V257.928C168.291 257.664 168.123 257.496 167.835 257.496H161.211ZM181.073 267.288C185.033 267.288 187.457 264.768 187.457 261.528C187.457 258.576 185.585 256.584 182.657 256.32L186.065 252.96C186.233 252.792 186.281 252.624 186.281 252.408V250.632C186.281 250.368 186.137 250.2 185.849 250.2H177.209C176.969 250.2 176.777 250.296 176.705 250.56L175.985 252.528C175.865 252.84 176.009 253.032 176.321 253.032H182.633L179.369 256.224C179.225 256.368 179.153 256.56 179.153 256.776V258.312C179.153 258.6 179.297 258.768 179.585 258.768H180.641C182.777 258.768 184.313 259.608 184.313 261.528C184.313 263.184 183.161 264.432 181.025 264.432C179.153 264.432 177.761 263.664 176.801 262.608C176.561 262.344 176.369 262.392 176.225 262.728L175.505 264.696C175.433 264.936 175.481 265.104 175.649 265.272C176.753 266.376 178.769 267.288 181.073 267.288Z"
-        fill="#3D3D3D"
-      />
-    </svg>
-  );
-};
