+++ client/packages/portal-client/src/components/portal-header/Header.tsx
@@ -1,8 +1,13 @@
 import { MenuButton, PortalHeader } from '@equinor/portal-ui';
+import { clearSelectedPhase } from '../../utils/phases/phases';
 
 export function Header() {
   return (
-    <PortalHeader MenuButton={MenuButton} title="Project Portal">
+    <PortalHeader
+      onLogoClick={clearSelectedPhase}
+      MenuButton={MenuButton}
+      title="Project Portal"
+    >
       Hello Project Portal
     </PortalHeader>
   );
