+++ client/packages/portal-client/src/utils/queryClient/index.ts
@@ -1,0 +1,1 @@
+export * from './queryClient';
