+++ client/packages/portal-client/src/utils/queryClient/queryClient.ts
@@ -1,0 +1,2 @@
+import { QueryClient } from 'react-query';
+export const queryClient = new QueryClient();
