+++ client/packages/portal-core/src/hooks/use-phases.ts
@@ -1,10 +1,9 @@
-import { Fusion } from '@equinor/fusion-framework-react';
-import { useFramework } from '@equinor/fusion-framework-react/hooks';
-import { WorkSurfaces } from '../types/portal-config';
+import { useQuery } from 'react-query';
+import { phases } from '../mock.ts/phases';
+import { Phase } from '../types/portal-config';
+/**
+ * Hook for getting phases from api
+ */
 export function usePhases() {
-  const { modules } = useFramework() as Fusion<
-    [{ name: 'phase'; initialize: () => { phases: WorkSurfaces[] } }]
-  >;
-
-  return modules.phase.phases as WorkSurfaces[];
+  return useQuery<Phase[]>(['Phases'], () => Promise.resolve(phases));
 }
