+++ client/packages/portal-core/src/mock.ts/phase-icons/Execution.tsx
@@ -1,0 +1,32 @@
+export const Execution = () => {
+  return (
+    <svg
+      width="305"
+      height="305"
+      viewBox="0 0 305 305"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M155.245 41.9414C155.245 39.9346 157.87 39.1781 158.938 40.8772L196.415 100.512C196.616 100.831 196.722 101.2 196.722 101.576V149C196.722 150.105 195.827 151 194.722 151H157.245C156.14 151 155.245 150.105 155.245 149L155.245 41.9414Z"
+        fill="#FF6670"
+      />
+      <mask id="path-2-inside-1_173_74322" fill="white">
+        <path
+          fillRule="evenodd"
+          clipRule="evenodd"
+          d="M196.722 101.576C196.722 101.2 196.616 100.831 196.415 100.512L158.938 40.8774C157.87 39.1782 155.245 39.9348 155.245 41.9416V149C155.245 150.105 156.14 151 157.245 151H194.722C195.827 151 196.722 150.105 196.722 149V101.576ZM226 157C226 155.896 225.105 155 224 155H81C79.8954 155 79 155.896 79 157V177C79 178.105 79.8954 179 81 179L99.5685 179C100.673 179 101.568 179.896 101.568 181V218C101.568 219.105 102.464 220 103.568 220H114.207C115.312 220 116.207 219.105 116.207 218V181C116.207 179.896 117.103 179 118.207 179H186.793C187.897 179 188.793 179.896 188.793 181V218C188.793 219.105 189.688 220 190.793 220H201.432C202.536 220 203.432 219.105 203.432 218V181C203.432 179.896 204.327 179 205.432 179L224 179C225.105 179 226 178.105 226 177V157ZM101 105C101 103.896 101.895 103 103 103H148C149.105 103 150 103.896 150 105V149C150 150.105 149.105 151 148 151H103C101.895 151 101 150.105 101 149V105ZM150.365 70.0001C150.365 68.8955 149.47 68.0001 148.365 68.0001H122.477C121.373 68.0001 120.477 68.8955 120.477 70.0001V97.0001C120.477 98.1047 121.373 99.0001 122.477 99.0001H148.365C149.47 99.0001 150.365 98.1047 150.365 97.0001V70.0001Z"
+        />
+      </mask>
+      <path
+        d="M224 179V175V179ZM205.432 179V183V179ZM99.5685 179L99.5685 175L99.5685 179ZM81 179L81 183L81 179ZM196.415 100.512L199.802 98.3838L196.415 100.512ZM155.551 43.0057L193.029 102.641L199.802 98.3838L162.325 38.749L155.551 43.0057ZM159.245 149V41.9416H151.245V149H159.245ZM194.722 147H157.245V155H194.722V147ZM192.722 101.576V149H200.722V101.576H192.722ZM81 159H224V151H81V159ZM83 177V157H75V177H83ZM99.5685 175L81 175L81 183L99.5685 183L99.5685 175ZM105.568 218V181H97.5685V218H105.568ZM114.207 216H103.568V224H114.207V216ZM112.207 181V218H120.207V181H112.207ZM186.793 175H118.207V183H186.793V175ZM192.793 218V181H184.793V218H192.793ZM201.432 216H190.793V224H201.432V216ZM199.432 181V218H207.432V181H199.432ZM224 175L205.432 175V183L224 183V175ZM222 157V177H230V157H222ZM148 99.0001H103V107H148V99.0001ZM154 149V105H146V149H154ZM103 155H148V147H103V155ZM97 105V149H105V105H97ZM122.477 72.0001H148.365V64.0001H122.477V72.0001ZM124.477 97.0001V70.0001H116.477V97.0001H124.477ZM148.365 95.0001H122.477V103H148.365V95.0001ZM146.365 70.0001V97.0001H154.365V70.0001H146.365ZM148.365 103C151.679 103 154.365 100.314 154.365 97.0001H146.365C146.365 95.8955 147.261 95.0001 148.365 95.0001V103ZM116.477 97.0001C116.477 100.314 119.163 103 122.477 103V95.0001C123.582 95.0001 124.477 95.8955 124.477 97.0001H116.477ZM148.365 72.0001C147.261 72.0001 146.365 71.1047 146.365 70.0001H154.365C154.365 66.6864 151.679 64.0001 148.365 64.0001V72.0001ZM122.477 64.0001C119.163 64.0001 116.477 66.6864 116.477 70.0001H124.477C124.477 71.1047 123.582 72.0001 122.477 72.0001V64.0001ZM103 147C104.105 147 105 147.896 105 149H97C97 152.314 99.6863 155 103 155V147ZM146 149C146 147.896 146.895 147 148 147V155C151.314 155 154 152.314 154 149H146ZM103 99.0001C99.6863 99.0001 97 101.686 97 105H105C105 106.105 104.105 107 103 107V99.0001ZM148 107C146.895 107 146 106.105 146 105H154C154 101.686 151.314 99.0001 148 99.0001V107ZM224 183C227.314 183 230 180.314 230 177H222C222 175.896 222.895 175 224 175V183ZM207.432 181C207.432 182.105 206.536 183 205.432 183V175C202.118 175 199.432 177.687 199.432 181H207.432ZM201.432 224C204.745 224 207.432 221.314 207.432 218H199.432C199.432 216.896 200.327 216 201.432 216V224ZM184.793 218C184.793 221.314 187.479 224 190.793 224V216C191.897 216 192.793 216.896 192.793 218H184.793ZM186.793 183C185.688 183 184.793 182.105 184.793 181H192.793C192.793 177.687 190.106 175 186.793 175V183ZM120.207 181C120.207 182.105 119.312 183 118.207 183V175C114.894 175 112.207 177.687 112.207 181H120.207ZM114.207 224C117.521 224 120.207 221.314 120.207 218H112.207C112.207 216.896 113.103 216 114.207 216V224ZM97.5685 218C97.5685 221.314 100.255 224 103.568 224V216C104.673 216 105.568 216.896 105.568 218H97.5685ZM99.5685 183C98.4639 183 97.5685 182.105 97.5685 181H105.568C105.568 177.687 102.882 175 99.5685 175L99.5685 183ZM75 177C75 180.314 77.6863 183 81 183L81 175C82.1046 175 83 175.896 83 177H75ZM224 159C222.895 159 222 158.105 222 157H230C230 153.687 227.314 151 224 151V159ZM81 151C77.6863 151 75 153.687 75 157H83C83 158.105 82.1046 159 81 159V151ZM194.722 155C198.036 155 200.722 152.314 200.722 149H192.722C192.722 147.896 193.617 147 194.722 147V155ZM151.245 149C151.245 152.314 153.931 155 157.245 155V147C158.349 147 159.245 147.896 159.245 149H151.245ZM193.029 102.641C192.828 102.322 192.722 101.953 192.722 101.576H200.722C200.722 100.447 200.403 99.3402 199.802 98.3838L193.029 102.641ZM162.325 38.749C159.122 33.6516 151.245 35.9212 151.245 41.9416H159.245C159.245 43.9483 156.619 44.7049 155.551 43.0057L162.325 38.749Z"
+        fill="#3D3D3D"
+        mask="url(#path-2-inside-1_173_74322)"
+      />
+      <path
+        d="M97.594 267C102.034 267 105.37 263.904 105.37 258.6C105.37 253.296 102.034 250.2 97.594 250.2H92.074C91.786 250.2 91.618 250.368 91.618 250.632V266.568C91.618 266.832 91.786 267 92.074 267H97.594ZM94.714 253.032H97.378C100.09 253.032 102.226 254.688 102.226 258.6C102.226 262.512 100.09 264.168 97.378 264.168H94.714V253.032ZM117.313 257.496C117.073 257.496 116.881 257.592 116.809 257.856L116.089 259.824C115.969 260.136 116.113 260.328 116.425 260.328H121.201C120.673 262.8 118.993 264.408 116.209 264.408C112.729 264.408 110.713 261.936 110.713 258.6C110.713 255.264 112.897 252.792 116.209 252.792C118.033 252.792 119.545 253.656 120.457 254.664C120.649 254.88 120.865 254.88 121.081 254.712L122.761 253.344C122.977 253.176 123.001 252.936 122.833 252.72C121.201 250.848 118.873 249.912 116.257 249.912C111.361 249.912 107.569 253.608 107.569 258.6C107.569 263.592 111.193 267.288 116.257 267.288C121.033 267.288 124.177 263.952 124.393 259.176V257.928C124.393 257.664 124.225 257.496 123.937 257.496H117.313ZM137.174 267.288C141.134 267.288 143.558 264.768 143.558 261.528C143.558 258.576 141.686 256.584 138.758 256.32L142.166 252.96C142.334 252.792 142.382 252.624 142.382 252.408V250.632C142.382 250.368 142.238 250.2 141.95 250.2H133.31C133.07 250.2 132.878 250.296 132.806 250.56L132.086 252.528C131.966 252.84 132.11 253.032 132.422 253.032H138.734L135.47 256.224C135.326 256.368 135.254 256.56 135.254 256.776V258.312C135.254 258.6 135.398 258.768 135.686 258.768H136.742C138.878 258.768 140.414 259.608 140.414 261.528C140.414 263.184 139.262 264.432 137.126 264.432C135.254 264.432 133.862 263.664 132.902 262.608C132.662 262.344 132.47 262.392 132.326 262.728L131.606 264.696C131.534 264.936 131.582 265.104 131.75 265.272C132.854 266.376 134.87 267.288 137.174 267.288ZM158.433 260.976C158.721 260.976 158.889 260.808 158.889 260.544V258.576C158.889 258.312 158.721 258.144 158.433 258.144H151.929C151.665 258.144 151.497 258.312 151.497 258.576V260.544C151.497 260.808 151.665 260.976 151.929 260.976H158.433ZM173.321 267C177.761 267 181.097 263.904 181.097 258.6C181.097 253.296 177.761 250.2 173.321 250.2H167.801C167.513 250.2 167.345 250.368 167.345 250.632V266.568C167.345 266.832 167.513 267 167.801 267H173.321ZM170.441 253.032H173.105C175.817 253.032 177.953 254.688 177.953 258.6C177.953 262.512 175.817 264.168 173.105 264.168H170.441V253.032ZM193.039 257.496C192.799 257.496 192.607 257.592 192.535 257.856L191.815 259.824C191.695 260.136 191.839 260.328 192.151 260.328H196.927C196.399 262.8 194.719 264.408 191.935 264.408C188.455 264.408 186.439 261.936 186.439 258.6C186.439 255.264 188.623 252.792 191.935 252.792C193.759 252.792 195.271 253.656 196.183 254.664C196.375 254.88 196.591 254.88 196.807 254.712L198.487 253.344C198.703 253.176 198.727 252.936 198.559 252.72C196.927 250.848 194.599 249.912 191.983 249.912C187.087 249.912 183.295 253.608 183.295 258.6C183.295 263.592 186.919 267.288 191.983 267.288C196.759 267.288 199.903 263.952 200.119 259.176V257.928C200.119 257.664 199.951 257.496 199.663 257.496H193.039ZM212.145 267C212.433 267 212.577 266.832 212.577 266.568V263.4H213.801C214.041 263.4 214.233 263.304 214.305 263.04L215.025 261.072C215.145 260.76 215.001 260.568 214.689 260.568H212.577V250.632C212.577 250.368 212.433 250.2 212.145 250.2H210.153C209.937 250.2 209.769 250.272 209.625 250.464L202.065 260.856C201.945 261.024 201.897 261.192 201.897 261.408V262.968C201.897 263.232 202.041 263.4 202.329 263.4H209.481V266.568C209.481 266.832 209.649 267 209.937 267H212.145ZM209.481 254.856V260.568H205.329L209.481 254.856Z"
+        fill="#3D3D3D"
+      />
+    </svg>
+  );
+};
