+++ client/packages/portal-core/src/mock.ts/phase-icons/Renewable.tsx
@@ -1,0 +1,26 @@
+export const Renewable = () => {
+  return (
+    <svg
+      width="305"
+      height="305"
+      viewBox="0 0 305 305"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        fillRule="evenodd"
+        clipRule="evenodd"
+        d="M155.852 98.7052L152.193 56.6302L148.643 97.4556L119.5 132.15L151.659 105.137L184.255 132.518L155.852 98.7052ZM209.508 156.362C210.294 157.298 209.035 158.557 208.099 157.771L156.3 114.26L167.692 230.589C169.79 232.641 171.87 234 175 234C179.25 234 181.564 231.493 184.617 228.151C187.569 224.92 191.166 221 197.75 221C201.006 221 203.413 221.959 205.362 223.438C207.034 224.708 208.338 226.353 209.483 227.798L209.661 228.022C209.722 228.099 209.784 228.176 209.844 228.252C211.137 229.872 212.33 231.263 213.917 232.289C215.454 233.281 217.477 234 220.5 234C221.605 234 222.5 234.895 222.5 236C222.5 237.105 221.605 238 220.5 238C216.773 238 213.984 237.094 211.747 235.649C209.561 234.237 208.019 232.378 206.718 230.748C206.613 230.616 206.51 230.487 206.409 230.359L206.388 230.334L206.387 230.332C205.181 228.816 204.191 227.572 202.943 226.624C201.68 225.666 200.119 225 197.75 225C193.084 225 190.556 227.58 187.57 230.849L187.445 230.987L187.296 231.15C184.466 234.255 181.053 238 175 238C168.947 238 165.534 234.255 162.704 231.15L162.539 230.968C162.502 230.928 162.466 230.889 162.43 230.849C159.444 227.58 156.916 225 152.25 225C147.55 225 144.765 227.625 141.667 230.879C141.501 231.054 141.333 231.232 141.163 231.412L141.162 231.413C138.334 234.404 134.936 238 129.5 238C123.852 238 119.546 234.13 116.114 230.972C115.936 230.807 115.76 230.645 115.585 230.484C113.986 229.008 112.551 227.684 111.097 226.674C109.504 225.567 108.109 225 106.75 225C105.473 225 104.481 225.499 103.39 226.503C102.595 227.234 101.904 228.069 101.084 229.06L101.083 229.06C100.639 229.597 100.156 230.18 99.5989 230.817C98.11 232.519 96.2802 234.342 93.752 235.724C91.2043 237.116 88.0578 238 84 238C82.8954 238 82 237.105 82 236C82 234.895 82.8954 234 84 234C87.4422 234 89.9207 233.259 91.834 232.214C93.7667 231.158 95.2337 229.731 96.5886 228.183C96.9708 227.746 97.3672 227.269 97.774 226.779L97.7741 226.779L97.7742 226.779C98.705 225.658 99.6898 224.472 100.681 223.56C102.238 222.126 104.152 221 106.75 221C109.266 221 111.465 222.058 113.38 223.389C115.077 224.568 116.724 226.09 118.282 227.529L118.282 227.53L118.283 227.53L118.644 227.864L118.823 228.028C122.454 231.37 125.648 234 129.5 234C132.077 234 133.959 232.831 135.948 230.983L147.754 113.641L95.6561 157.403C94.72 158.189 93.4609 156.93 94.2472 155.994L144.768 95.8503L151.197 21.9134C151.303 20.6955 153.084 20.6955 153.19 21.9134L159.728 97.0999L209.508 156.362ZM140.408 226.445C143.234 223.677 146.805 221 152.25 221C157.321 221 160.621 223.326 163.218 225.871L152 111.391L140.408 226.445Z"
+        fill="#3D3D3D"
+      />
+      <path
+        d="M152 111.392L166.042 232C164.194 229.075 158.7 223.225 151.5 223.225C144.3 223.225 139.468 229.12 137.952 232.067L152 111.392Z"
+        fill="#FF6670"
+      />
+      <path
+        d="M97.594 267C102.034 267 105.37 263.904 105.37 258.6C105.37 253.296 102.034 250.2 97.594 250.2H92.074C91.786 250.2 91.618 250.368 91.618 250.632V266.568C91.618 266.832 91.786 267 92.074 267H97.594ZM94.714 253.032H97.378C100.09 253.032 102.226 254.688 102.226 258.6C102.226 262.512 100.09 264.168 97.378 264.168H94.714V253.032ZM117.313 257.496C117.073 257.496 116.881 257.592 116.809 257.856L116.089 259.824C115.969 260.136 116.113 260.328 116.425 260.328H121.201C120.673 262.8 118.993 264.408 116.209 264.408C112.729 264.408 110.713 261.936 110.713 258.6C110.713 255.264 112.897 252.792 116.209 252.792C118.033 252.792 119.545 253.656 120.457 254.664C120.649 254.88 120.865 254.88 121.081 254.712L122.761 253.344C122.977 253.176 123.001 252.936 122.833 252.72C121.201 250.848 118.873 249.912 116.257 249.912C111.361 249.912 107.569 253.608 107.569 258.6C107.569 263.592 111.193 267.288 116.257 267.288C121.033 267.288 124.177 263.952 124.393 259.176V257.928C124.393 257.664 124.225 257.496 123.937 257.496H117.313ZM137.174 267.288C141.134 267.288 143.558 264.768 143.558 261.528C143.558 258.576 141.686 256.584 138.758 256.32L142.166 252.96C142.334 252.792 142.382 252.624 142.382 252.408V250.632C142.382 250.368 142.238 250.2 141.95 250.2H133.31C133.07 250.2 132.878 250.296 132.806 250.56L132.086 252.528C131.966 252.84 132.11 253.032 132.422 253.032H138.734L135.47 256.224C135.326 256.368 135.254 256.56 135.254 256.776V258.312C135.254 258.6 135.398 258.768 135.686 258.768H136.742C138.878 258.768 140.414 259.608 140.414 261.528C140.414 263.184 139.262 264.432 137.126 264.432C135.254 264.432 133.862 263.664 132.902 262.608C132.662 262.344 132.47 262.392 132.326 262.728L131.606 264.696C131.534 264.936 131.582 265.104 131.75 265.272C132.854 266.376 134.87 267.288 137.174 267.288ZM158.433 260.976C158.721 260.976 158.889 260.808 158.889 260.544V258.576C158.889 258.312 158.721 258.144 158.433 258.144H151.929C151.665 258.144 151.497 258.312 151.497 258.576V260.544C151.497 260.808 151.665 260.976 151.929 260.976H158.433ZM173.321 267C177.761 267 181.097 263.904 181.097 258.6C181.097 253.296 177.761 250.2 173.321 250.2H167.801C167.513 250.2 167.345 250.368 167.345 250.632V266.568C167.345 266.832 167.513 267 167.801 267H173.321ZM170.441 253.032H173.105C175.817 253.032 177.953 254.688 177.953 258.6C177.953 262.512 175.817 264.168 173.105 264.168H170.441V253.032ZM193.039 257.496C192.799 257.496 192.607 257.592 192.535 257.856L191.815 259.824C191.695 260.136 191.839 260.328 192.151 260.328H196.927C196.399 262.8 194.719 264.408 191.935 264.408C188.455 264.408 186.439 261.936 186.439 258.6C186.439 255.264 188.623 252.792 191.935 252.792C193.759 252.792 195.271 253.656 196.183 254.664C196.375 254.88 196.591 254.88 196.807 254.712L198.487 253.344C198.703 253.176 198.727 252.936 198.559 252.72C196.927 250.848 194.599 249.912 191.983 249.912C187.087 249.912 183.295 253.608 183.295 258.6C183.295 263.592 186.919 267.288 191.983 267.288C196.759 267.288 199.903 263.952 200.119 259.176V257.928C200.119 257.664 199.951 257.496 199.663 257.496H193.039ZM212.145 267C212.433 267 212.577 266.832 212.577 266.568V263.4H213.801C214.041 263.4 214.233 263.304 214.305 263.04L215.025 261.072C215.145 260.76 215.001 260.568 214.689 260.568H212.577V250.632C212.577 250.368 212.433 250.2 212.145 250.2H210.153C209.937 250.2 209.769 250.272 209.625 250.464L202.065 260.856C201.945 261.024 201.897 261.192 201.897 261.408V262.968C201.897 263.232 202.041 263.4 202.329 263.4H209.481V266.568C209.481 266.832 209.649 267 209.937 267H212.145ZM209.481 254.856V260.568H205.329L209.481 254.856Z"
+        fill="#3D3D3D"
+      />
+    </svg>
+  );
+};
