+++ client/packages/portal-ui/src/lib/portal-header/PortalHeader.tsx
@@ -7,6 +7,7 @@ interface PortalHeaderProps {
   Navigation?: React.FC;
   MenuButton: React.FC;
   title: string;
+  onLogoClick: () => void;
 }
 
 export function PortalHeader({
@@ -14,12 +15,13 @@ export function PortalHeader({
   Navigation,
   MenuButton,
   title,
+  onLogoClick,
 }: React.PropsWithChildren<PortalHeaderProps>): JSX.Element {
   return (
     <StyledTopBar>
       <StyledHeader>
         <MenuButton />
-        <PortalLogo title={title} />
+        <PortalLogo title={title} onClick={onLogoClick} />
         {Navigation && <Navigation />}
       </StyledHeader>
