+++ client/packages/portal-utils/src/rxjs/hooks/index.ts
@@ -1,0 +1,1 @@
+export * from './useObservable';
