+++ client/packages/portal-utils/src/rxjs/index.ts
@@ -1,0 +1,3 @@
+export * from './observableProxy';
+export * from './hooks';
+export * from './types';
