+++ client/packages/portal-utils/src/rxjs/types/index.ts
@@ -1,0 +1,1 @@
+export * from './observableProxy';
